+++ src/modules/Elsa.Workflows.Management/Features/WorkflowManagementFeature.cs
@@ -226,6 +226,7 @@ public class WorkflowManagementFeature(IModule module) : FeatureBase(module)
             .AddMemoryStore<WorkflowInstance, MemoryWorkflowInstanceStore>()
             .AddActivityProvider<TypedActivityProvider>()
             .AddActivityProvider<WorkflowDefinitionActivityProvider>()
+            .AddScoped<WorkflowDefinitionActivityDescriptorFactory>()
             .AddScoped<WorkflowDefinitionActivityProvider>()
             .AddScoped<IWorkflowDefinitionActivityRegistryUpdater, WorkflowDefinitionActivityRegistryUpdater>()
             .AddScoped<IWorkflowDefinitionService, WorkflowDefinitionService>()
